commit lol

+++ app/shared/functions/authentication/session.ts
@@ -5,7 +5,9 @@ import { getSessionId } from "./cookie";
 import { UNAUTHORIZED } from "./errors";
 
 export const getCurrentUserClient = async () => {
-  const { id } = getSessionId();
+  const payload = getSessionId();
+  if (typeof payload === "string") return UNAUTHORIZED;
+  const { id } = payload;
   const session = await db.session.findFirst({
     where: { id },
     include: { user: { include: { client: true } } },
@@ -14,7 +16,22 @@ export const getCurrentUserClient = async () => {
   return session.user;
 };
 
+export const getCurrentUserStaff = async () => {
+  const payload = getSessionId();
+  if (typeof payload === "string") return UNAUTHORIZED;
+  const { id } = payload;
+  const session = await db.session.findFirst({
+    where: { id },
+    include: { user: { include: { staff: true } } },
+  });
+  if (!session) return UNAUTHORIZED;
+  if (session.role === "CLIENT") return UNAUTHORIZED;
+  return session.user;
+};
+
 export const deleteCurrentSession = async () => {
-  const { id } = getSessionId();
+  const payload = getSessionId();
+  if (typeof payload === "string") return UNAUTHORIZED;
+  const { id } = payload;
   await db.session.delete({ where: { id } });
 };
